fix(greater_): preserve float64 precision instead of converting to float32
The original implementation converted inputs to float32 using x.to(tl.float32),
which caused precision loss and incorrect comparison results for float64 inputs.

Issue:
- Values like 16777216.5 and 16777216.4 are distinguishable in float64
- But both round to 16777216.0 in float32
- This caused greater_ to return wrong results for float64 inputs

Fix:
- Remove the .to(tl.float32) conversion
- Let Triton use the input's native dtype for comparison
- This preserves full precision for float64

Testing:
- Added test_greater_fp64_precision to verify float64 precision is preserved
- Test uses values that expose the precision loss bug if present

+++ src/flag_gems/ops/greater_.py
@@ -28,13 +28,13 @@ logger = logging.getLogger(__name__)
 @pointwise_dynamic(promotion_methods=[(0, 1, "ALWAYS_BOOL")])
 @triton.jit
 def greater__inplace_func(x, y):
-    return x.to(tl.float32) > y
+    return x > y
 
 
 @pointwise_dynamic(is_tensor=[True, False], promotion_methods=[(0, 1, "ALWAYS_BOOL")])
 @triton.jit
 def greater__scalar_inplace_func(x, y):
-    return x.to(tl.float32) > y
+    return x > y
 
 
 def greater_(A, B):
